fix(rust): signed locals for negative-capable (int) casts of doubles (#160)
C truncates a negative double to a negative int; Rust's f64->usize cast
saturates it to 0. Locals assigned a whole-RHS (int)(float) cast are now
classified signed (i32) — real_vars threaded into the classifier so plain
double locals are seen regardless of name — and their assignments render
'as i32'. sqrt/fabs/abs inners stay usize (provably non-negative; HMA).
Positions the backend cannot render sign-faithfully (casts nested in
arithmetic, compound assigns, conditions) now FAIL GENERATION loudly
instead of silently saturating. Deliberately not transitive through
var-to-var copies (would churn the extremum family's hot loops).

MAVP: the dual-role size/period temp is split in the input so the size
arithmetic stays usize (an i32 endIdx cast silently truncated >=2^31
inputs — caught by review, now pinned by a crate-wide 'index domain never
narrows to i32' test). MAVP + the HT family (DCPeriodInt flips i32,
C-faithful) verified bitwise against the frozen oracle and cross-language.

synth2/ added to the synthetic gate: negative casts in both use classes
(magnitude clamps + two's-complement masking), bit-exact in all four
languages, batch + stream. synth1 restaged to the supported whole-RHS
cast form, exercising the i32/usize mixing paths.

+++ ta_codegen/output/rust/library/src/ta_func/ht_sine.rs
@@ -247,7 +247,7 @@ impl Core {
         let mut smoothPrice_Idx: usize = 0;
         let mut maxIdx_smoothPrice: usize = 49;
         let mut idx: usize = 0_usize;
-        let mut DCPeriodInt: usize = 0_usize;
+        let mut DCPeriodInt: i32 = 0_i32;
         let mut DCPhase: f64 = 0.0_f64;
         let mut DCPeriod: f64 = 0.0_f64;
         let mut imagPart: f64 = 0.0_f64;
@@ -526,15 +526,15 @@ impl Core {
             smoothPeriod = (0.67 as f64).mul_add(smoothPeriod, 0.33 * period);
             // Compute Dominant Cycle Phase
             DCPeriod = smoothPeriod + 0.5;
-            DCPeriodInt = (DCPeriod as usize) as usize;
+            DCPeriodInt = (DCPeriod) as i32;
             realPart = 0.0;
             imagPart = 0.0;
             // idx is used to iterate for up to 50 of the last
             // value of smoothPrice.
             idx = smoothPrice_Idx;
-            // for( i = 0; i < DCPeriodInt; i += 1 )
+            // for( i = 0; ((i) as i32) < DCPeriodInt; i += 1 )
             i = 0;
-            while i < DCPeriodInt {
+            while ((i) as i32) < DCPeriodInt {
                 tempReal = (i as f64) * constDeg2RadBy360 / (DCPeriodInt as f64);
                 tempReal2 = smoothPrice[idx];
                 realPart += (tempReal).sin() * tempReal2;
@@ -689,7 +689,7 @@ impl Core {
         let mut smoothPrice_Idx: usize = 0;
         let mut maxIdx_smoothPrice: usize = 49;
         let mut idx: usize = 0_usize;
-        let mut DCPeriodInt: usize = 0_usize;
+        let mut DCPeriodInt: i32 = 0_i32;
         let mut DCPhase: f64 = 0.0_f64;
         let mut DCPeriod: f64 = 0.0_f64;
         let mut imagPart: f64 = 0.0_f64;
@@ -923,13 +923,13 @@ impl Core {
             period = (0.2 as f64).mul_add(period, 0.8 * tempReal);
             smoothPeriod = (0.67 as f64).mul_add(smoothPeriod, 0.33 * period);
             DCPeriod = smoothPeriod + 0.5;
-            DCPeriodInt = (DCPeriod as usize) as usize;
+            DCPeriodInt = (DCPeriod) as i32;
             realPart = 0.0;
             imagPart = 0.0;
             idx = smoothPrice_Idx;
-            // for( i = 0; i < DCPeriodInt; i += 1 )
+            // for( i = 0; ((i) as i32) < DCPeriodInt; i += 1 )
             i = 0;
-            while i < DCPeriodInt {
+            while ((i) as i32) < DCPeriodInt {
                 tempReal = (i as f64) * constDeg2RadBy360 / (DCPeriodInt as f64);
                 tempReal2 = smoothPrice[idx];
                 realPart += (tempReal).sin() * tempReal2;
@@ -1043,7 +1043,7 @@ struct HtSineStreamState {
     constDeg2RadBy360: f64,
     smoothPeriod: f64,
     idx: usize,
-    DCPeriodInt: usize,
+    DCPeriodInt: i32,
     DCPhase: f64,
     DCPeriod: f64,
     imagPart: f64,
@@ -1206,15 +1206,15 @@ impl Core {
         sp.smoothPeriod = (0.67 as f64).mul_add(sp.smoothPeriod, 0.33 * sp.period);
         // Compute Dominant Cycle Phase
         sp.DCPeriod = sp.smoothPeriod + 0.5;
-        sp.DCPeriodInt = (sp.DCPeriod as usize) as usize;
+        sp.DCPeriodInt = (sp.DCPeriod) as i32;
         sp.realPart = 0.0;
         sp.imagPart = 0.0;
         // idx is used to iterate for up to 50 of the last
         // value of smoothPrice.
         sp.idx = sp.smoothPrice_Idx;
-        // for( sp.i = 0; sp.i < sp.DCPeriodInt; sp.i += 1 )
+        // for( sp.i = 0; ((sp.i) as i32) < sp.DCPeriodInt; sp.i += 1 )
         sp.i = 0;
-        while sp.i < sp.DCPeriodInt {
+        while ((sp.i) as i32) < sp.DCPeriodInt {
             sp.tempReal = (sp.i as f64) * sp.constDeg2RadBy360 / (sp.DCPeriodInt as f64);
             sp.tempReal2 = sp.cb_smoothPrice[sp.idx];
             sp.realPart += (sp.tempReal).sin() * sp.tempReal2;
@@ -1341,7 +1341,7 @@ impl Core {
         let mut smoothPrice_Idx: usize = 0;
         let mut maxIdx_smoothPrice: usize = 49;
         let mut idx: usize = 0_usize;
-        let mut DCPeriodInt: usize = 0_usize;
+        let mut DCPeriodInt: i32 = 0_i32;
         let mut DCPhase: f64 = 0.0_f64;
         let mut DCPeriod: f64 = 0.0_f64;
         let mut imagPart: f64 = 0.0_f64;
@@ -1620,15 +1620,15 @@ impl Core {
             smoothPeriod = (0.67 as f64).mul_add(smoothPeriod, 0.33 * period);
             // Compute Dominant Cycle Phase
             DCPeriod = smoothPeriod + 0.5;
-            DCPeriodInt = (DCPeriod as usize) as usize;
+            DCPeriodInt = (DCPeriod) as i32;
             realPart = 0.0;
             imagPart = 0.0;
             // idx is used to iterate for up to 50 of the last
             // value of smoothPrice.
             idx = smoothPrice_Idx;
-            // for( i = 0; i < DCPeriodInt; i += 1 )
+            // for( i = 0; ((i) as i32) < DCPeriodInt; i += 1 )
             i = 0;
-            while i < DCPeriodInt {
+            while ((i) as i32) < DCPeriodInt {
                 tempReal = (i as f64) * constDeg2RadBy360 / (DCPeriodInt as f64);
                 tempReal2 = smoothPrice[idx];
                 realPart += (tempReal).sin() * tempReal2;
@@ -1865,7 +1865,7 @@ impl Core {
         let mut smoothPrice_Idx: usize = 0;
         let mut maxIdx_smoothPrice: usize = 49;
         let mut idx: usize = 0_usize;
-        let mut DCPeriodInt: usize = 0_usize;
+        let mut DCPeriodInt: i32 = 0_i32;
         let mut DCPhase: f64 = 0.0_f64;
         let mut DCPeriod: f64 = 0.0_f64;
         let mut imagPart: f64 = 0.0_f64;
@@ -2144,15 +2144,15 @@ impl Core {
             smoothPeriod = (0.67 as f64).mul_add(smoothPeriod, 0.33 * period);
             // Compute Dominant Cycle Phase
             DCPeriod = smoothPeriod + 0.5;
-            DCPeriodInt = (DCPeriod as usize) as usize;
+            DCPeriodInt = (DCPeriod) as i32;
             realPart = 0.0;
             imagPart = 0.0;
             // idx is used to iterate for up to 50 of the last
             // value of smoothPrice.
             idx = smoothPrice_Idx;
-            // for( i = 0; i < DCPeriodInt; i += 1 )
+            // for( i = 0; ((i) as i32) < DCPeriodInt; i += 1 )
             i = 0;
-            while i < DCPeriodInt {
+            while ((i) as i32) < DCPeriodInt {
                 tempReal = (i as f64) * constDeg2RadBy360 / (DCPeriodInt as f64);
                 tempReal2 = smoothPrice[idx];
                 realPart += (tempReal).sin() * tempReal2;
